requirements.txt
python-telegram-bot[job-queue]~=21.10
aiosqlite~=0.20.0
python-dotenv~=1.1.0
